pyproject.toml
[project]
name = "cloudrun-otel-collector-python-sample-app"
version = "0.1.0"
description = "A sample FastAPI application with OpenTelemetry Collector for Cloud Run"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "grpcio>=1.76.0",
    "loguru>=0.7.3",
    "opentelemetry-exporter-otlp>=1.39.0",
    "opentelemetry-exporter-otlp-proto-http>=1.39.0",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "opentelemetry-instrumentation-requests>=0.60b0",
    "opentelemetry-sdk>=1.39.0",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn>=0.38.0",
]
